chore: renaming module
